ia/requirements.txt
tensorflow-hub==0.12.0
flask
pandas